backend/pyproject.toml
[tool.poetry]
name = "lukalibre-backend"
version = "0.1.0"
description = "Backend para la plataforma de educación financiera LukaLibre"
authors = ["LukaLibre Team <contacto@lukalibre.org>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
asyncpg = "^0.28.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.19"}
alembic = "^1.11.1"
pydantic = {extras = ["email"], version = "^2.1.1"}
pydantic-settings = "^2.0.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
authlib = "^1.2.1"
httpx = "^0.24.1"
rich = "^13.4.2"
structlog = "^23.1.0"
slowapi = "^0.1.8"
redis = {extras = ["hiredis"], version = "^4.6.0"}
orjson = "^3.9.2"
psycopg2-binary = "^2.9.10"
itsdangerous = "^2.2.0"
google-auth = "^2.39.0"
google-api-python-client = "^2.167.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
openai = "^1.76.0"
aiofiles = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.4.1"
flake8 = "^6.0.0"
httpx = "^0.24.1"
faker = "^19.2.0"
asgi-lifespan = "^2.1.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101", "B104"]
